requirements.txt
celery==5.0.2
django-hashid-field==3.1.3
django-model-utils==4.0.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
django-celery-beat==2.1.0
django-celery-results==1.2.1
Django==3.1.3
drf-spectacular==0.10.0
python-decouple==3.3
psycopg2-binary==2.8.6
redis==3.5.3
sentry-sdk==0.19.2
sendgrid==6.4.6
django-pgcrypto==2.0.0
